+++ package/model/lib/model.dart
@@ -15,6 +15,6 @@ export 'src/feature/loading/loading.dart';
 export 'src/feature/push_notification/push_notification.dart';
 export 'src/feature/quiz/quiz.dart';
 export 'src/feature/quiz_result/quiz_result.dart';
-export 'src/feature/search_condition/search_condition.dart';
+export 'src/feature/search_condition/exporter.dart';
 export 'src/feature/user_info/exporter.dart';
 export 'src/util/util.dart';
