+++ package/model/lib/src/feature/analytics/application/analytics_service.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../../search_condition/search_condition.dart';
+import '../../search_condition/exporter.dart';
 
 part 'analytics_service.g.dart';
