+++ package/model/lib/src/feature/quiz_result/application/quiz_result_service.dart
@@ -4,7 +4,7 @@ import 'package:riverpod_annotation/riverpod_annotation.dart';
 import '../../auth/infrastructure/auth_repository.dart';
 import '../../daily_quiz/application/daily_quiz_state.dart';
 import '../../quiz/domain/quiz_state.dart';
-import '../../search_condition/application/search_condition_state.dart';
+import '../../search_condition/search_condition_state.dart';
 import '../infrastructure/quiz_result_repository.dart';
 import 'quiz_result_state.dart';
