+++ package/model/lib/src/feature/search_condition/search_condition.freezed.dart
