+++ package/model/lib/src/feature/search_condition/search_condition_constant.dart
@@ -1,5 +1,5 @@
-import '../../../util/enum/hitting_stats_type.dart';
-import '../domain/search_condition.dart';
+import '../../util/enum/hitting_stats_type.dart';
+import 'search_condition.dart';
 
 final defaultSearchCondition = SearchCondition(
   teamList: [
