+++ package/model/lib/src/feature/search_condition/search_condition_state.dart
@@ -1,7 +1,7 @@
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-import '../domain/search_condition.dart';
-import '../infrastructure/search_condition_repository.dart';
+import 'search_condition.dart';
+import 'search_condition_repository.dart';
 
 part 'search_condition_state.g.dart';
