+++ package/model/test/dummy_data/dummy_hitter.dart
@@ -1,5 +1,5 @@
 import 'package:model/src/feature/quiz/domain/quiz.dart';
-import 'package:model/src/feature/search_condition/domain/search_condition.dart';
+import 'package:model/src/feature/search_condition/search_condition.dart';
 import 'package:model/src/util/enum/hitting_stats_type.dart';
 
 const dummySearchCondition = SearchCondition(
