+++ package/model/test/dummy_data/dummy_search_condition.dart
@@ -1,4 +1,4 @@
-import 'package:model/src/feature/search_condition/domain/search_condition.dart';
+import 'package:model/src/feature/search_condition/search_condition.dart';
 import 'package:model/src/util/enum/hitting_stats_type.dart';
 
 /// 1つの球団を選択しているダミーデータ
